refactor: add changes to profiles router, users router to pass the tests

+++ src/routes/profiles/index.ts
@@ -6,9 +6,7 @@ import type { ProfileEntity } from '../../utils/DB/entities/DBProfiles';
 const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
   fastify
 ): Promise<void> => {
-  fastify.get('/', async function (request, reply): Promise<
-    ProfileEntity[]
-  > {
+  fastify.get('/', async function (request, reply): Promise<ProfileEntity[]> {
     return await fastify.db.profiles.findMany();
   });
 
@@ -26,7 +24,7 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
         equals: id,
       });
       if (Object.is(profile, null)) {
-        reply.notFound("Profile doesn't exist")
+        reply.notFound("Profile doesn't exist");
       }
       return profile as ProfileEntity;
     }
@@ -41,6 +39,19 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
     },
     async function (request, reply): Promise<ProfileEntity> {
       const body = request.body;
+      const { userId, memberTypeId } = body;
+
+      const users = await fastify.db.users.findMany();
+      const profiles = await fastify.db.profiles.findMany();
+      const memberTypes = await fastify.db.memberTypes.findMany();
+
+      const isUserExist = users.find((u) => u.id === userId);
+      const isProfileExist = profiles.find((p) => p.userId === userId);
+      const isMemberTypeExist = memberTypes.find((m) => m.id === memberTypeId);
+
+      if (!isUserExist || isProfileExist || !isMemberTypeExist) {
+        throw fastify.httpErrors.badRequest();
+      }
       return await fastify.db.profiles.create(body);
     }
   );
@@ -54,6 +65,13 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
     },
     async function (request, reply): Promise<ProfileEntity> {
       const id = request.params.id;
+      const profile = await fastify.db.profiles.findOne({
+        key: 'id',
+        equals: id,
+      });
+      if (Object.is(profile, null)) {
+        throw fastify.httpErrors.badRequest();
+      }
       return await fastify.db.profiles.delete(id);
     }
   );
@@ -67,9 +85,15 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
-      const id = request.params.id;
+      const id: string = request.params.id;
       const body = request.body;
-      const profile = await fastify.db.profiles.findOne({key:'id', equals:id}) as ProfileEntity;
+      const profile = (await fastify.db.profiles.findOne({
+        key: 'id',
+        equals: id,
+      })) as ProfileEntity;
+      if (Object.is(profile, null)) {
+        throw fastify.httpErrors.badRequest();
+      }
       return await fastify.db.profiles.change(id, Object.assign(profile, body));
     }
   );
